+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.14
 
+ - Removed horizontal support for `<Card />` (DOM structure issues)
  - Removed default props from OverlayTrigger
 
 ## 1.0.13